+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.1.0 
 
 - Enable line-by-line verse layout in addition to paragraph layout.
+- Auto-play next chapter when streaming (#62)
 
 ## 1.0.1 - April 3, 2026